# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dongjing server",
        "title": "gin-base-framework Server api",
        "termsOfService": "https://github.com/hexiaoyun128",
        "contact": {},
        "license": {},
        "version": "2.0"
    },
    "paths": {
...
        "/auth/menu/{id}": {
            "get": {
                "description": "菜单获得",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "菜单获得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单获得ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ]
            },
            "put": {
                "description": "菜单修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "菜单修改",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "轮播图组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否有效",
                        "name": "active",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "time",
                        "description": "开始时间",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "time",
                        "description": "结束时间",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "轮播组名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "轮播时间间隔(s)",
                        "name": "display_time",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"id\":1,\"parent_id\":0,\"name\":\"测试修改\",\"route\":\"route-system\",\"component\":\"SystemComponent\",\"icon\":\"system-icon\",\"sequence\":1},\"message\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/menutree": {
            "get": {
                "description": "系统菜单树,系统管理员才可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "系统菜单树",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":[{\"id\":1,\"parent_id\":0,\"name\":\"系统设置\",\"route\":\"route-system\",\"component\":\"SystemComponent\",\"icon\":\"system-icon\",\"sequence\":1,\"children\":[{\"id\":2,\"parent_id\":1,\"name\":\"系统用户\",\"route\":\"/system/user\",\"icon\":\"system-user\",\"sequence\":1}]},{\"id\":3,\"parent_id\":0,\"name\":\"分类模块\",\"route\":\"route_modyle\",\"component\":\"Module\",\"icon\":\"module\",\"sequence\":2,\"children\":[{\"id\":4,\"parent_id\":3,\"name\":\"模块列表\",\"route\":\"/auth/category\",\"icon\":\"category\",\"sequence\":1}]},{\"id\":5,\"parent_id\":0,\"name\":\"公司管理\",\"route\":\"route_company\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2,\"children\":[{\"id\":6,\"parent_id\":5,\"name\":\"公司分类\",\"route\":\"route_company_1\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2,\"children\":[{\"id\":7,\"parent_id\":6,\"name\":\"一级公司\",\"route\":\"route_company_2\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2},{\"id\":8,\"parent_id\":6,\"name\":\"二级公司\",\"route\":\"route_company_3\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2}]}]}],\"message\":\"success\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/auth/usermenu": {
            "get": {
                "description": "用户菜单树获得",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "用户菜单树获得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":[{\"id\":1,\"parent_id\":0,\"name\":\"系统设置\",\"route\":\"route-system\",\"component\":\"SystemComponent\",\"icon\":\"system-icon\",\"sequence\":1,\"children\":[{\"id\":2,\"parent_id\":1,\"name\":\"系统用户\",\"route\":\"/system/user\",\"icon\":\"system-user\",\"sequence\":1}]},{\"id\":3,\"parent_id\":0,\"name\":\"分类模块\",\"route\":\"route_modyle\",\"component\":\"Module\",\"icon\":\"module\",\"sequence\":2,\"children\":[{\"id\":4,\"parent_id\":3,\"name\":\"模块列表\",\"route\":\"/auth/category\",\"icon\":\"category\",\"sequence\":1}]},{\"id\":5,\"parent_id\":0,\"name\":\"公司管理\",\"route\":\"route_company\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2,\"children\":[{\"id\":6,\"parent_id\":5,\"name\":\"公司分类\",\"route\":\"route_company_1\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2,\"children\":[{\"id\":7,\"parent_id\":6,\"name\":\"一级公司\",\"route\":\"route_company_2\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2},{\"id\":8,\"parent_id\":6,\"name\":\"二级公司\",\"route\":\"route_company_3\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2}]}]}],\"message\":\"success\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/usermenutree/{user_id}": {
            "get": {
                "description": "用户菜单树,系统管理员才可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "用户菜单树",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":[{\"id\":1,\"parent_id\":0,\"name\":\"系统设置\",\"route\":\"route-system\",\"component\":\"SystemComponent\",\"icon\":\"system-icon\",\"sequence\":1,\"children\":[{\"id\":2,\"parent_id\":1,\"name\":\"系统用户\",\"route\":\"/system/user\",\"icon\":\"system-user\",\"sequence\":1}]},{\"id\":3,\"parent_id\":0,\"name\":\"分类模块\",\"route\":\"route_modyle\",\"component\":\"Module\",\"icon\":\"module\",\"sequence\":2,\"children\":[{\"id\":4,\"parent_id\":3,\"name\":\"模块列表\",\"route\":\"/auth/category\",\"icon\":\"category\",\"sequence\":1}]},{\"id\":5,\"parent_id\":0,\"name\":\"公司管理\",\"route\":\"route_company\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2,\"children\":[{\"id\":6,\"parent_id\":5,\"name\":\"公司分类\",\"route\":\"route_company_1\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2,\"children\":[{\"id\":7,\"parent_id\":6,\"name\":\"一级公司\",\"route\":\"route_company_2\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2},{\"id\":8,\"parent_id\":6,\"name\":\"二级公司\",\"route\":\"route_company_3\",\"component\":\"Company\",\"icon\":\"company\",\"sequence\":2}]}]}],\"message\":\"success\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}